refactor: 클래스명 통일 및 디렉토리 그룹핑
