+++ app/src/main/java/com/example/guru2/allergy/AllergyActivity.kt
@@ -1,10 +1,11 @@
-package com.example.guru2
+package com.example.guru2.allergy
 
 import android.os.Bundle
 import android.view.View
 import android.widget.ArrayAdapter
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.example.guru2.R
 import com.example.guru2.databinding.ActivityAllergyBinding
 import com.google.android.material.chip.Chip
