+++ app/src/main/java/com/example/guru2/db/DatabaseHelper.kt
@@ -1,9 +1,10 @@
-package com.example.guru2
+package com.example.guru2.db
 
 import android.content.ContentValues
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
+import com.example.guru2.fridge.Ingredient
 
 class DatabaseHelper(context: Context) : SQLiteOpenHelper(context, DATABASE_NAME, null, DATABASE_VERSION) {
