+++ app/src/main/java/com/example/guru2/recipe/RecipeActivity.kt
@@ -1,13 +1,11 @@
-package com.example.guru2
+package com.example.guru2.recipe
 
-import android.graphics.Color
 import android.os.Bundle
-import android.widget.RadioButton
-import android.widget.RadioGroup
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.example.guru2.R
 
 class RecipeActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
