+++ app/src/main/res/layout/activity_ask_recipe_sample.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AskRecipeActivity">
+    tools:context=".recipe.AskRecipeActivity">
 
     <TextView
         android:layout_width="wrap_content"
