+++ app/src/main/res/layout/activity_recipe1.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".recipe1">
+    tools:context=".recipe.recipe1">
 
     <!-- 타이틀 텍스트뷰 -->
     <TextView
