+++ app/src/main/res/layout/list_item_allergy.xml
@@ -7,10 +7,11 @@
     android:background="@drawable/item_bg"
     android:orientation="vertical">
 
-<TextView
-android:id="@+id/textItem"
-android:layout_width="wrap_content"
-android:layout_height="wrap_content"
-android:textSize="14sp"
-android:textColor="#000000" />
-    </LinearLayout>
+    <TextView
+        android:id="@+id/textItem"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textSize="14sp"
+        android:textColor="#000000" />
+
+</LinearLayout>
